+++ feature/page_two/src/main/java/ru/msokolov/onlineshop/page_two/data/entity/DetailedInfoEntity.kt
@@ -1,4 +1,4 @@
-package ru.msokolov.onlineshop.page_two.presentation.ui.data.entity
+package ru.msokolov.onlineshop.page_two.data.entity
 
 data class DetailedInfoEntity(
     val name: String,
