+++ feature/page_two/src/main/java/ru/msokolov/onlineshop/page_two/di/PageTwoComponent.kt
@@ -1,0 +1,2 @@
+package ru.msokolov.onlineshop.page_two.di
+
