+++ feature/page_two/src/main/java/ru/msokolov/onlineshop/page_two/domain/repository/DetailedInfoRepository.kt
@@ -1,0 +1,8 @@
+package ru.msokolov.onlineshop.page_two.domain.repository
+
+import ru.msokolov.onlineshop.page_two.data.entity.DetailedInfoEntity
+
+interface DetailedInfoRepository {
+
+    suspend fun getDetailedInfoData() : DetailedInfoEntity
+}
