+++ feature/page_two/src/main/java/ru/msokolov/onlineshop/page_two/domain/usecase/GetDetailedInfoUseCase.kt
@@ -1,6 +1,6 @@
-package ru.msokolov.onlineshop.page_two.presentation.ui.domain.usecase
+package ru.msokolov.onlineshop.page_two.domain.usecase
 
-import ru.msokolov.onlineshop.page_two.presentation.ui.domain.repository.DetailedInfoRepository
+import ru.msokolov.onlineshop.page_two.domain.repository.DetailedInfoRepository
 import javax.inject.Inject
 
 class GetDetailedInfoUseCase @Inject constructor(private val repository: DetailedInfoRepository) {
