--- feature/page_two/src/main/java/ru/msokolov/onlineshop/page_two/presentation/ui/di/PageTwoComponent.kt
@@ -1,2 +1,0 @@
-package ru.msokolov.onlineshop.page_two.presentation.ui.di
-
