--- feature/page_two/src/main/java/ru/msokolov/onlineshop/page_two/presentation/ui/domain/repository/DetailedInfoRepository.kt
@@ -1,8 +1,0 @@
-package ru.msokolov.onlineshop.page_two.presentation.ui.domain.repository
-
-import ru.msokolov.onlineshop.page_two.presentation.ui.data.entity.DetailedInfoEntity
-
-interface DetailedInfoRepository {
-
-    suspend fun getDetailedInfoData() : DetailedInfoEntity
-}
